+++ app/core/fastapi_depends.py
@@ -8,7 +8,10 @@ from app.repositories.document_embeddings_repository import (
 )
 from app.services.content_rag_service import ContentRagService
 from app.services.content_service import ContentService
+from app.services.exam_rag_service import ExamRagService
 from app.services.exam_service import ExamService
+from app.services.mindmap_rag_service import MindmapRagService
+from app.services.slide_rag_service import SlideRagService
 
 
 def get_logger():
@@ -35,6 +38,21 @@ def get_content_rag_service(request: Request) -> ContentRagService:
     return request.app.state.content_rag_service
 
 
+def get_slide_rag_service(request: Request) -> SlideRagService:
+    """Get the slide rag service."""
+    return request.app.state.slide_rag_service
+
+
+def get_mindmap_rag_service(request: Request) -> MindmapRagService:
+    """Get the mindmap rag service."""
+    return request.app.state.mindmap_rag_service
+
+
+def get_exam_rag_service(request: Request) -> ExamRagService:
+    """Get the exam rag service."""
+    return request.app.state.exam_rag_service
+
+
 ExamServiceDep = Annotated[ExamService, Depends(get_exam_service)]
 ContentRagServiceDep = Annotated[
     ContentRagService, Depends(get_content_rag_service)
@@ -43,3 +61,10 @@ ContentServiceDep = Annotated[ContentService, Depends(get_content_service)]
 DocumentEmbeddingsRepositoryDep = Annotated[
     DocumentEmbeddingsRepository, Depends(get_doc_repository)
 ]
+
+# Specialized RAG service dependencies
+SlideRagServiceDep = Annotated[SlideRagService, Depends(get_slide_rag_service)]
+MindmapRagServiceDep = Annotated[
+    MindmapRagService, Depends(get_mindmap_rag_service)
+]
+ExamRagServiceDep = Annotated[ExamRagService, Depends(get_exam_rag_service)]
